requirements.txt
click==8.1.7
fastapi==0.109.2
httpx==0.26.0
lxml==5.1.0
python-dotenv==1.0.1
SQLAlchemy==2.0.25
sqlmodel==0.0.14
starlette==0.36.3
typing-extensions==4.9.0
uvicorn==0.27.0.post1
